Add files via upload
Customer Segmentation is very important before launching the products and for validating purposes. Once we understand the Customer Segment then we could launch smoothly and keep our product on top. This Data Science Project would help the company to identify the Customer segment and assist to launch a better product.